+++ packages/contact-center/cc-components/src/components/UserState/user-state.tsx
@@ -6,7 +6,8 @@ import {formatTime} from '../../utils';
 import './user-state.scss';
 import {SelectNext, Text} from '@momentum-ui/react-collaboration';
 import {Item} from '@react-stately/collections';
-import {Icon} from '@momentum-design/components/dist/react';
+import {Icon, Tooltip} from '@momentum-design/components/dist/react';
+import {userStateLabels} from './constant';
 
 const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
   const {
@@ -76,6 +77,20 @@ const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
     }
   };
 
+  const getTooltipText = () => {
+    if (customState && customState.developerName === 'ENGAGED') {
+      const currentStateObj = idleCodes.find((item) => item.id === currentState);
+
+      if (currentStateObj.name === AgentUserState.Available) {
+        return userStateLabels.customWithAvailableTooltip;
+      } else {
+        return userStateLabels.customWithIdleStateTooltip.replace(/{{.*?}}/g, currentStateObj.name);
+      }
+    }
+
+    return userStateLabels.availableTooltip;
+  };
+
   // Sorts the dropdown items by keeping 'Available' at the top and sorting the rest alphabetically by name
   const sortedItems = [
     ...items.filter((item) => item.name === AgentUserState.Available),
@@ -83,8 +98,9 @@ const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
   ];
 
   return (
-    <div className="user-state-container">
+    <div className="user-state-container" data-testid="user-state-container">
       <SelectNext
+        id="user-state-tooltip"
         label=""
         aria-label="user-state"
         direction="bottom"
@@ -96,6 +112,7 @@ const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
         selectedKey={selectedKey}
         items={sortedItems}
         className={`state-select ${getDropdownClass()}`}
+        data-testid="state-select"
       >
         {(item) => {
           const isRonaOrEngaged = [AgentUserState.RONA, AgentUserState.Engaged].includes(
@@ -104,14 +121,18 @@ const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
           const shouldHighlight = currentState === item.id || (isRonaOrEngaged && item.id === previousSelectableState);
 
           return (
-            <Item key={item.id} textValue={item.name}>
-              <div className={`item-container ${shouldHighlight ? `selected ${getIconStyle(item).class}` : ''}`}>
+            <Item key={item.id} textValue={item.name} data-testid={`state-item-${item.name}`}>
+              <div
+                className="item-container"
+                data-testid={`item-container ${shouldHighlight ? `selected ${getIconStyle(item).class}` : ''}`}
+              >
                 <Icon
                   name={getIconStyle(item).iconName}
                   title=""
                   className={`state-icon ${getIconStyle(item).class}`}
+                  data-testid="state-icon"
                 />
-                <Text className={`state-name ${getIconStyle(item).class}`} tagName={'small'}>
+                <Text className={`state-name ${getIconStyle(item).class}`} tagName={'small'} data-testid="state-name">
                   {item.name}
                 </Text>
               </div>
@@ -120,13 +141,22 @@ const UserStateComponent: React.FunctionComponent<IUserState> = (props) => {
         }}
       </SelectNext>
 
+      <Tooltip placement="bottom" color="contrast" delay="0, 0" className="tooltip" triggerID="user-state-tooltip">
+        <Text tagName="small" className="tooltip-text">
+          {getTooltipText()}
+        </Text>
+      </Tooltip>
+
       {!customState && (
-        <span className={`elapsedTime ${isSettingAgentStatus ? 'elapsedTime-disabled' : ''}`}>
+        <span
+          className={`elapsedTime ${isSettingAgentStatus ? 'elapsedTime-disabled' : ''}`}
+          data-testid="elapsed-time"
+        >
           {lastIdleStateChangeElapsedTime >= 0 ? formatTime(lastIdleStateChangeElapsedTime) + ' / ' : ''}
           {formatTime(elapsedTime)}
         </span>
       )}
-      <Icon className="select-arrow-icon" name="arrow-down-bold" title="" />
+      <Icon className="select-arrow-icon" name="arrow-down-bold" title="" data-testid="select-arrow-icon" />
     </div>
   );
 };